Moved volume 4 2025 release changes to public flutter widgets repository

+++ packages/syncfusion_flutter_barcodes/lib/src/barcode_generator/renderers/one_dimensional/ean13_renderer.dart
@@ -88,7 +88,7 @@ class EAN13Renderer extends SymbologyRenderer {
       ratio = symbology!.module!.toDouble();
     } else {
       // Calculates the bar length based on number of individual bar codes
-      final int singleModule = (width ~/ barTotalLength).toInt();
+      final int singleModule = width ~/ barTotalLength;
       ratio = singleModule.toDouble();
       final double leftPadding = (width - (barTotalLength * ratio)) / 2;
       left += leftPadding;
