+++ packages/syncfusion_flutter_barcodes/lib/src/barcode_generator/renderers/one_dimensional/upce_renderer.dart
@@ -84,7 +84,7 @@ class UPCERenderer extends SymbologyRenderer {
       ratio = symbology!.module!.toDouble();
     } else {
       //Calculates the bar length based on number of individual bar codes
-      final int singleModule = (width ~/ barTotalLength).toInt();
+      final int singleModule = width ~/ barTotalLength;
       ratio = singleModule.toDouble();
       final double leftPadding = (width - (barTotalLength * ratio)) / 2;
       left += leftPadding;
